feat: use jwt token on login
