+++ server/src/router.ts
@@ -1,7 +1,7 @@
 import {Router} from 'express';
 const router = Router();
 import * as controller from './mvc/controllers/controllers';
-import registerUser from './mvc/controllers/signup';
+import registerUser, { loginUser } from './mvc/controllers/auth';
 
 router.get('/', controller.getAllTutors);
 router.get('/:id', controller.getTutor);
@@ -10,5 +10,6 @@ router.put('/:id', controller.updateTutor);
 router.delete('/:id', controller.deleteTutor);
 
 router.post('/signup', registerUser);
+router.post('/login', loginUser);
 
 export default router;
